--- mlir/include/mlir/Dialect/MIOpen/utility/common.h
@@ -1,5 +1,0 @@
-#ifndef COMMON_H
-#define COMMON_H
-
-
-#endif
